Need to fix Navbar
the color on the nav bar is different and the background of the nav bar doesn't fade to black when it scrolls down. #Hint, think I found a clue in the "js-scroll-trigger" 
Also, the scroll should be slow and it isn't. It's jumping down. May need to make nav all over again. 
Also, added a bootstrap file. Need to eliminate some unnecessary files in the folder.